Initial sketch 160-bit (32+128) RawVal

+++ soroban-env-host/src/native_contract/token/error.rs
@@ -21,6 +21,6 @@ pub enum ContractError {
 
 impl From<ContractError> for Status {
     fn from(err: ContractError) -> Self {
-        Status::from_contract_error(err as u32)
+        Status::from_contract_error(err as u64)
     }
 }
